--- a/openquake/calculators/event_based_risk.py
+++ b/openquake/calculators/event_based_risk.py
@@ -223,13 +223,11 @@ def event_based_risk(df, crmodel, monitor):
                             int(oq.asset_correlation))
 
     with monitor('reading assets', measuremem=True):
-        dfs = []
-        for id0taxo, s0, s1 in monitor.read('start-stop'):
-            ass = monitor.read('assets', slice(s0, s1)).set_index('ordinal')
-            country = crmodel.countries[id0taxo // TWO24]
-            dfs.append((ass, country))
-    outgen = output_gen(df, dfs, crmodel, rng, monitor)
-    del dfs
+        # saving a lot of memory with respect to a simple read('assets')
+        cols = monitor.read_pdcolumns('assets').split()
+        assdic = {col: monitor.read('assets/' + col) for col in cols}
+    outgen = output_gen(df, assdic, crmodel, rng, monitor)
+    del assdic
     with monitor('aggregating losses', measuremem=True) as agg_mon:
         avg, alt = aggreg(outgen, crmodel, ARK, aggids, rlz_id, oq.ideduc,
                           monitor)
@@ -241,9 +239,10 @@ def event_based_risk(df, crmodel, monitor):
                 out_bytes=out_bytes)
 
 
-def output_gen(df, dfs, crmodel, rng, monitor):
+def output_gen(df, assdic, crmodel, rng, monitor):
     """
     :param df: GMF dataframe (a slice of events)
+    :param assdic: a dictionary column name -> array
     :param crmodel: CompositeRiskModel instance
     :param rng: random number generator
     :param monitor: Monitor instance
@@ -253,7 +252,10 @@ def output_gen(df, dfs, crmodel, rng, monitor):
     fil_mon = monitor('filtering GMFs', measuremem=False)
     sids = df.sid.to_numpy()
     monitor.ctime = 0
-    for adf, country in dfs:
+    for id0taxo, s0, s1 in monitor.read('start-stop'):
+        adf = pandas.DataFrame({col: assdic[col][s0:s1] for col in assdic})
+        adf = adf.set_index('ordinal')
+        country = crmodel.countries[id0taxo // TWO24]
         with fil_mon:
             # *crucial* for the performance of the next step
             gmf_df = df[numpy.isin(sids, adf.site_id.unique())]