This is for Pakistan small on engine192 (6x speedup in read assets):
```
# before
| calc_1295, maxmem=376.5 GB   | time_sec  | memory_mb | counts  |
|------------------------------+-----------+-----------+---------|
| total ebr_from_gmfs          | 77_717    | 725.7     | 752     |
| computing risk               | 37_652    | 0.0       | 423_335 |
| aggregating losses           | 24_768    | 1_626     | 382     |
| reading assets               | 13_725    | 1_439     | 382     |
| reading crmodel              | 882.1     | 301.8     | 370     |
| EventBasedRiskCalculator.run | 723.2     | 1_947     | 1       |

# after
| calc_1294, maxmem=366.7 GB   | time_sec  | memory_mb | counts  |
|------------------------------+-----------+-----------+---------|
| total ebr_from_gmfs          | 67_730    | 509.7     | 752     |
| computing risk               | 38_380    | 0.0       | 423_335 |
| aggregating losses           | 25_648    | 983.2     | 382     |
| reading assets               | 2_103     | 1_701     | 382     |
| reading crmodel              | 899.2     | 301.7     | 370     |
| EventBasedRiskCalculator.run | 670.3     | 1_932     | 1       |
```